// ui/src/generated/api/model/dtoSetPermissions.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export interface DtoSetPermissions {
  delete: boolean;
  edit: boolean;
  editItems: boolean;
  manageAccess: boolean;
  publish: boolean;
  transfer: boolean;
  view: boolean;
  viewAccess: boolean;
}